requirements.txt
# 工具合规扫描 Agent 服务 - Python 依赖
# Python 3.9+

# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP 客户端（用于 AI API 调用和联网搜索）
httpx==0.25.1
requests==2.31.0

# 配置管理
pyyaml==6.0.1
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0

# 数据库
sqlalchemy>=2.0.23
aiosqlite==0.19.0

# 日志
loguru==0.7.2

# 环境变量管理
python-dotenv==1.0.0

# 工具库
python-multipart==0.0.6  # FastAPI 文件上传支持

# 开发依赖（可选）
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
